Refactoring GC api
